RIPEMD: Remove the old RIPEMD.c module due to copyright concerns
The copyright notice and disclaimer at the top of RIPEMD.c says:

    Copyright (c) Katholieke Universiteit Leuven 1996, All Rights Reserved
    The Katholieke Universiteit Leuven makes no representations concerning
    either the merchantability of this software or the suitability of this
    software for any particular purpose. It is provided "as is" without
    express or implied warranty of any kind. These notices must be retained
    in any copies of any part of this documentation and/or software.

To me, it seems as though the software was intended to be freely distributed,
as long as the notice and disclaimer are included.  However, no such
permission is explicitly granted, so I am removing the module (without looking
at its contents).